add feature to 1.x: make permission of docs configurable
By default markdown files in the docs folder are imported (or updated on
change in filesystem) into the database (example "release-notes") as
'protcted' (everyone - even without login can read, nobody can write).

This commit tries to make this configurable so that custom permissions
can be set. (For example if you want to restrict viewing the release
notes - and therefore the knowlege of the currently used HedgeDoc
version - to logged in users you can set the permission to 'protected'.)

Example configuration for config.json:
```
"specificPublicDocsPermissions": {
    "release-notes": "protected"
},
```

The default behaviour does not change and notes from the docs path
will be set to 'locked' if not otherwise specified.

  Signed-off-by: Andreas Boesen <boesen@belwue.de>